Make the incident history table into a component.
